// src/types/finance.ts
export interface FinanceData {
  nome: string;
  idade: string;

  filhos: number;

  renda: number;

  beneficio: string;

  objetivo: string;

  moradia: number;

  alimentacao: number;

  terapias: number;

  medicamentos: number;

  transporte: number;

  escola: number;

  outros: number;

  meta: number;

  prazo: string;
}